ml/requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # Para Parquet

# Base de datos
psycopg2-binary>=2.9.0   # PostgreSQL
sqlalchemy>=2.0.0
redis>=5.0.0             # Online store

# Validación
pydantic>=2.0.0
great-expectations>=1.0.0

# Feature Store
feast>=0.38.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0

# Experiment Tracking
mlflow>=2.10.0

# Utilidades
python-dotenv>=1.0.0
loguru>=0.7.0
click>=8.1.0             # CLI